+++ packages/atomic/src/components.d.ts
@@ -9,6 +9,7 @@ import { Engine, LogLevel, Result, ResultTemplateCondition } from "@coveo/headle
 import { Bindings } from "./utils/initialization-utils";
 import { i18n } from "i18next";
 import { InitializationOptions } from "./components/atomic-search-interface/atomic-search-interface";
+import { BaseFacetSearchResult } from "@coveo/headless/dist/api/search/facet-search/base/base-facet-search-response";
 export namespace Components {
     interface AtomicBreadcrumbManager {
         "categoryDivider": string;
@@ -127,17 +128,14 @@ export namespace Components {
         "_id": string;
         /**
           * Whether the submit button should be placed before the input
-          * @default false
          */
         "leadingSubmitButton": boolean;
         /**
           * Maximum number of suggestions to display
-          * @default 5
          */
         "numberOfSuggestions": number;
         /**
           * The placeholder for the search box input
-          * @default ''
          */
         "placeholder": string;
     }
@@ -177,6 +175,19 @@ export namespace Components {
           * String key value
          */
         "value": string;
+    }
+    interface BaseFacet {
+        "hasActiveValues": boolean;
+        "label": string;
+    }
+    interface FacetSearch {
+        "facetSearchResults": BaseFacetSearchResult[];
+        "moreValuesAvailable": boolean;
+    }
+    interface FacetValue {
+        "isSelected": boolean;
+        "label": string;
+        "numberOfResults": number;
     }
 }
 declare global {
@@ -348,6 +359,24 @@ declare global {
         prototype: HTMLAtomicTextElement;
         new (): HTMLAtomicTextElement;
     };
+    interface HTMLBaseFacetElement extends Components.BaseFacet, HTMLStencilElement {
+    }
+    var HTMLBaseFacetElement: {
+        prototype: HTMLBaseFacetElement;
+        new (): HTMLBaseFacetElement;
+    };
+    interface HTMLFacetSearchElement extends Components.FacetSearch, HTMLStencilElement {
+    }
+    var HTMLFacetSearchElement: {
+        prototype: HTMLFacetSearchElement;
+        new (): HTMLFacetSearchElement;
+    };
+    interface HTMLFacetValueElement extends Components.FacetValue, HTMLStencilElement {
+    }
+    var HTMLFacetValueElement: {
+        prototype: HTMLFacetValueElement;
+        new (): HTMLFacetValueElement;
+    };
     interface HTMLElementTagNameMap {
         "atomic-breadcrumb-manager": HTMLAtomicBreadcrumbManagerElement;
         "atomic-category-facet": HTMLAtomicCategoryFacetElement;
@@ -377,6 +406,9 @@ declare global {
         "atomic-sort-dropdown": HTMLAtomicSortDropdownElement;
         "atomic-tab": HTMLAtomicTabElement;
         "atomic-text": HTMLAtomicTextElement;
+        "base-facet": HTMLBaseFacetElement;
+        "facet-search": HTMLFacetSearchElement;
+        "facet-value": HTMLFacetValueElement;
     }
 }
 declare namespace LocalJSX {
@@ -494,17 +526,14 @@ declare namespace LocalJSX {
         "_id"?: string;
         /**
           * Whether the submit button should be placed before the input
-          * @default false
          */
         "leadingSubmitButton"?: boolean;
         /**
           * Maximum number of suggestions to display
-          * @default 5
          */
         "numberOfSuggestions"?: number;
         /**
           * The placeholder for the search box input
-          * @default ''
          */
         "placeholder"?: string;
     }
@@ -543,6 +572,24 @@ declare namespace LocalJSX {
          */
         "value": string;
     }
+    interface BaseFacet {
+        "hasActiveValues": boolean;
+        "label": string;
+        "onDeselectAll"?: (event: CustomEvent<void>) => void;
+    }
+    interface FacetSearch {
+        "facetSearchResults": BaseFacetSearchResult[];
+        "moreValuesAvailable": boolean;
+        "onFacetSearch"?: (event: CustomEvent<string>) => void;
+        "onResultSelected"?: (event: CustomEvent<BaseFacetSearchResult>) => void;
+        "onShowMoreResults"?: (event: CustomEvent<void>) => void;
+    }
+    interface FacetValue {
+        "isSelected": boolean;
+        "label": string;
+        "numberOfResults": number;
+        "onFacetValueSelected"?: (event: CustomEvent<void>) => void;
+    }
     interface IntrinsicElements {
         "atomic-breadcrumb-manager": AtomicBreadcrumbManager;
         "atomic-category-facet": AtomicCategoryFacet;
@@ -572,6 +619,9 @@ declare namespace LocalJSX {
         "atomic-sort-dropdown": AtomicSortDropdown;
         "atomic-tab": AtomicTab;
         "atomic-text": AtomicText;
+        "base-facet": BaseFacet;
+        "facet-search": FacetSearch;
+        "facet-value": FacetValue;
     }
 }
 export { LocalJSX as JSX };
@@ -606,6 +656,9 @@ declare module "@stencil/core" {
             "atomic-sort-dropdown": LocalJSX.AtomicSortDropdown & JSXBase.HTMLAttributes<HTMLAtomicSortDropdownElement>;
             "atomic-tab": LocalJSX.AtomicTab & JSXBase.HTMLAttributes<HTMLAtomicTabElement>;
             "atomic-text": LocalJSX.AtomicText & JSXBase.HTMLAttributes<HTMLAtomicTextElement>;
+            "base-facet": LocalJSX.BaseFacet & JSXBase.HTMLAttributes<HTMLBaseFacetElement>;
+            "facet-search": LocalJSX.FacetSearch & JSXBase.HTMLAttributes<HTMLFacetSearchElement>;
+            "facet-value": LocalJSX.FacetValue & JSXBase.HTMLAttributes<HTMLFacetValueElement>;
         }
     }
 }